Started refactoring before working on global abstraction.

+++ GlobalFinanceManager/Source/EntryIterator.cpp
@@ -129,7 +129,7 @@ EntryIterator& EntryIterator::operator--()
 		return *this;
 	}
 	
-	for (size_t i = current_index_ - 1; i >= 0; i--) {
+	for (int i = current_index_ - 1; i >= 0; i--) {
 		if (request_->IsValid((*target_container_p_)[i])) {
 			current_index_ = i;
 			return *this;
@@ -172,7 +172,7 @@ EntryIterator EntryIterator::operator--(int)
 	
 	EntryIterator previous_state(*this);
 
-	for (size_t i = current_index_ - 1; i >= 0; i--) {
+	for (int i = current_index_ - 1; i >= 0; i--) {
 		if (request_->IsValid((*target_container_p_)[i])) {
 			current_index_ = i;
 			return previous_state;
@@ -224,7 +224,7 @@ EntryIterator& EntryIterator::ToEnd()
 		return *this;
 	}
 	
-	for (size_t i = target_container_p_->size() - 1; i >= 0; i--)
+	for (int i = target_container_p_->size() - 1; i >= 0; i--)
 	{
 		if (request_->IsValid((*target_container_p_)[i])) {
 			current_index_ = i + 1;
